cleanup setup.py a little
