+++ datacube_ows/wsgi.py
@@ -10,3 +10,10 @@ os.environ.setdefault("DATACUBE_CONFIG_PATH", "/opt/odc/.datacube.conf.local")
 
 from datacube_ows.ogc import app
 application = app
+
+def main():
+    app.run()
+
+
+if __name__ == '__main__':
+    main()
