+++ setup.py
@@ -10,7 +10,27 @@ with open('HISTORY.rst') as history_file:
     history = history_file.read()
 
 requirements = [
-    'datacube', 'flask'
+    'datacube',
+    'Flask',
+    'flask_log_request_id',
+    'requests',
+    'affine',
+    'click',
+    'colour',
+    'lxml',
+    'matplotlib',
+    'numpy',
+    'opencensus',
+    'Pillow',
+    'prometheus_client',
+    'psycopg2',
+    'python_dateutil',
+    'pytz',
+    'rasterio',
+    'regex',
+    'scikit-image',
+    'timezonefinderL',
+    'xarray'
 ]
 
 test_requirements = [
@@ -28,7 +48,8 @@ setup(
     url='https://github.com/opendatacube/datacube-wms',
     entry_points={
         'console_scripts': [
-            'datacube-ows=datacube_ows.wms_wsgi.__main__:main'
+            'datacube-ows=datacube_ows.wsgi:main',
+            'datacube-ows-update=datacube_ows.update_ranges:main'
         ]
     },
     packages=find_packages(),
